Decouple get command and query
